feat: implement ChannelState trait for various channel states
ChannelState is a trait that all the states of the lifecycle must implement
to present a common API regardless of which stage of the lifecycle we're in.

This means that it can only expose the data that's common to all states,
i.e. channel_id and role. We might also add in some pubkey info later if
needed.

The states are going to need to be serializable, so adding in some to_
and from_hex utility functions for keys as well.

Refactor the lifecycle test to make it easier to add tests that cover the
non-happy paths.