feat: persisting channel state (#19)
* feat: persisting channel state

Adds the `StateStore` trait which lets us create various implementations
for persisting state.

`FileStore` is a simple file-based implementation that stores the
channel state in a file with the same name.

Making `ChannelLifeCycle` derive `Deserialize`required touching a lot of
files, but it's the most robust solution, I feel.

A nice test that stores and loads the state for every transition on the
happy path illustrates that this works.

Be aware: Secret keys are stored in plain text in the current
implementation. We'll want to come up with a more secure implementation
in future.

* fix: review comments