feat: monero multig traits (#30)
Provides the correct (or very close to correct) traits for a multisig
wallet.

Specifically, the

* MultiSigWallet trait, which provides pure local wallet
functionality, and

* MultiSigService, which is expected to know how to
interact with the EventLoop for communicating with peers.

This allows multiple concrete implementations without having to recode a
lot of things, in particular
* A stubbed version for testing
* A version using Monero wallet RPC
* A version using Serai's FROST protocol?

Many types related to Monero transactions are simply stubbed out at this
time, and we expect to move them and flesh them out once the RPC
implementation is complete.

- All functions are async, so we can hook it up into an RPC
- Completes the Dummy impl, which stubs out all the methods
- Removes the old placeholder traits which are no longer required