feat: kes and funding tx placeholders
This commit completes the establish_new_channel method by adding in the
mechanisms to create a KES and the Funding Transaction.

Because this is an event-based system, we can't just call into the
various methods. Instead, we create TODO list items that are queued up
and handled in the event loop.