feat: funding transaction events
This PR adds the funding transaction events to the state machine and
stores the tx id hashes in the state for later use.

Potentially, both the merchant and customer have to provide funding to
the multisig-wallet, so the state is expanded to allow for that.

Also if the initial balance is zero, a funding transaction is not
required, so that case is also handled.