wip: big refactor and cleanup
* The state machine is far leaner and simpler.

Each stage is a simple struct with defined methods and a `next` method
to try transition to next state.

The state only holds information, and doesn't perform any work.

The PaymentChannel struct is updated to drive the new state machine,
including a test of the happy path.

* Reduce boilerplate

Replace lots of boilerplate with some macros that handle the common
pattern of the GreaseRequest-GreaseResponse pattern.

* funding tx watcher

Add a watch only wallet and simple dummy delegate impl to use it to
watch for the funding tx.

* Updates multisig examples

Copies the payment channel walkthrough to `payment_channel_demo.rs`
Updates the examples to use the more ergonomic `Curve25519*` structs.

* Scanning for funding tx works

Added a transaction scanner to DummyDelegate that can watch the
blockchain for a funding transaction.

To keep things simple for now WE ASSUME THAT ONLY THE CUSTOMER FUNDS THE
CHANNEL. This assumption can be dropped at a later stage.

This is included in the channel establishment flow.

* Merchant watches for fundign transaction.

Even though the customer sends the funding transaction, the merchant
side now correctly watches for the funding transaction as well.

The CLI tool will also rescan the blockchain when loading a channel
state from disk, if necessary.