feat: mvp multisig wallet
This PR provides a MVP for a multisig wallet, using
* MuSig for 2-of-2 signature aggregation
* Serai's FROST protocol for transaction construction.

Functionality is increadibly basic right now.
A wallet can scan for, and save UTXOs that it owns. But it does not
remove them when spent.

2 wallet can co-ordinate to create a new Transaction.
2 round-trips are needed:
- Generate and share preprocessing information.
- Generate and share signature shares.

Then either wallet can
- Consolidate and broadcast

See `multisig.rs` for a working example (assumes you have a node running
with some funds already in the multisig address).