chore: remove unused code (#51)
Quite a lot of dead code has accumulated to this point, so this is as
good a time as any to ditch it.

* remove todo list item

The async queue in server.rs is not needed right now. It might be in
future, but for now, get rid of it.

* remove use of key manager.

The multisig wallet keys are used as the canonical keys for each
channel, so the keymanager generation in the proposal is not required,
and confuses matters greatly.