feat: share initial proofs
Provides the P2P tooling and state management for the witness0
generation and proof exchange with the merchant from the customer side,
and vice-versa.

Also added a bunch of tests for MoneroAmount since we added the Sum
trait and there were no tests.

Switched FundingTransaction tracking in Establishing state to a HashMap
since funding tx's were being double-counted on rescans. This fixes
that.