feat: handle channel close request
* Copy state from open to closing

Copies the relevant state from Open to Closing state. Handles case where
no updates have been made, but we want to close the channel.

* feat: introduce closing addresses

A missing so far has been the specification of the closing addresses,
specifically the Monero addresses into which the funds will be sent when
the channel is closed.

This is now added, and the return addresses are committed to in the
channel id (though not yet in the channel proofs).

* add open -> closing transition

The p2p communication required to enter the co-operative channel closing
state is provided.

The ChannelClose event is amended to accept the current witness value
from the peer, which is needed to complete signing the final
transaction.

* Handle channel close request-reponse

Adds the logic from each peer side to request and respond to a channel
close event.

* use a consistent address for Bob

Changes BOB_ADDRESS to a new address that is used consistently
throughout.

add `refund_address` to global config file, so that the CLI doesn't have
to prompt the user all the time.