feat: Include monero transactions in updates (#60)
Inlcudes data to produce a valid on-chain transaction along with every
update. Making this happen entailed adding another round of
communication to the update phase.

Since I want updates to be atomic, i.e. act as if nothing happened if
the 2nd round trip doesn't happen, I added a new type `PendingUpdate`
that the Server holds onto to carry data between rounds, without
impacting the state (or having to roll it back on a failure).

feat: serialize preprocess data

Use the internal machinery of Serai to convert the Preprocess data into
a byte stream, so that it can be sent over the wire.

feat: add adapter signature logic

Adds the adaptor signature logic from the payment channel demo into
MultisigWallet.

The multisig wallet test is updated to make us of adapter signatures.

The update flow for channel updates uses the new adapter signature methods
in MultisigWallet.

fix and test

Dummy delegate needs to return valid Ed25519 scalars now in order for
the adapter signatires to work.

Fixed a few small bugs -- we were accidentally sending 0 as the delta
for every update. But the sanity checks caught that, so easy to track
down.