feat: broadcast final transaction (#61)
The logic for publishing the closing transaction to the blockchain is
added.

In order for the multisig wallet to be reconstructed deterministically,
we need to provide a seeded RNG to the `prepare` function.

We use the hash of the spend key for this purpose. Will need to check
that this is a secure approach.

Otherwise, end to end tests along the jhappy path work now!