Cucumber funding tx (#68)
* feat: simple wallet

Adds a single-signer wallet implementation. This is primarily used in
Cucumber tests, but could also be used for "one-shot" wallets in cases
where a spend key and associated UTXO is known and it needs to be sent
somewhere.

* cargo fix

* tests: fuding transaction in channels

Adds the ability to submit the funding transaction to a channel.

One change that's needed to is paramterise the polling interval on the
transaction watcher service. It was set up to be 5s (fine in real life
mainnet), but in tests, block interval times are sub 1ms, so we need to
watch much more frequently.

The EventHandlerOptions struct was introduced to cater for this
flexibility.

The User's monero address and view keys are derived from the secert key
rather than being hard-coded.

A bug in multisig wallet was fixed where different RNGs were being
passed to the wallet at different stages of the multisig wallet
creation.

* review comments