feat: cucumber channel payment (#69)
Adds steps to the cucumber tests allowing channels payments

`When Alice pays 0.1 XMR to Bob`

will initiate a payment of 0.1 XMR

You can also send multiple payments (for stress tests):

`When Alice pays 0.1 XMR to Bob 100 times`

You can also test refunds:

`When Bob refunds 0.1 XMR to Alice`

And you can check the channel status:

```gherkin
Then the channel balance is
      | customer_balance  |  1.15  |
      | merchant_balance  |  0.1  |
```

`And the transaction count is 11`.

-- A new utility function in `NetworkServer` has been added:
`transaction_count()`, which returns the number of updates the channel
has seen.

The variable tracking this has been moved to `ChannelMetadata` since it
is tightly bound to the current balance, and is useful to know
regardless of the lifecycle stage of the channel.