feat: preparing for noir-delegate
- Refactor the delegate module to make room for NoirDelegate without
  cluttering the main module.
  - Extract traits into their own module
  - Extract DelegateError into its own module
- Bring in circuits code from PR #67
- Bring in Helper functions and circuit code from #67

Right now the code cannot be integrated into Delegate traits is not
possible since there are input variables that the ZK circuits need that
are not provided by the trait functions (i.e. PublicInputs).

The specification documentation has been updated to reflect this new
need.

The shell calls have been modified so that the tests can pass locally
(and should also pass CI if the ENVars are correctly configured).