test: test vectors for babyjubjub
This commit adds a few fixes to the Scalar implementation,
adds a large number of tests to the suite, including an implmenmtation
comparison with the JS version of @zk-kit/baby-jubjub, which can be
generated using the included script (create_vectors).

PrimeFieldBits is implemnted for babyjubjub::Scalar.

We should now have a contant-time, zeroized version of BabyJubJub that
is compatible with Serai; DLEQ proofs in particular.