feat: serai-compatible babyjubjub implementation (#82)
We should now have a constant-time, zeroized version of BabyJubJub that is compatible with Serai (with the serai feature flag); DLEQ proofs in particular.

====
Features
====

The Ark libraries provide a very convenient way to implement specific instances of common elliptic curves (e.g. in Twisted Edward form), along with a large suite of standardised tests.

    Added BabyJubJub curve support: field types, constants, point types, arithmetic, and serialization for points and scalars.
    Integration with an external cipher-suite compatibility layer for scalar and group operations.

====
Serai compatibility (with the serai feature flag)
====

    New types, BjjPoint and Scalar are defined that are Zeroize and support constant-time operations.
    Field, PrimeField and PrimeFieldBits is implemented for babyjubjub::Scalar.
    Group and GroupEncoding is implemented for BjjPoint

====
Tests
====
A large number of tests are added to the suite. The generic test suite of Ark-FF and serai are included as well as a side-by-side implementation comparison of ECC operations with the JS version of @zk-kit/baby-jubjub.

The test vectors can be generated using the included script (create_vectors.mjs).