feat: open channel protocol (#90)
* feat: open channel protocol

Delivers a set of traits that prove a generic implementation of the
Channel Opening Protocol that works for any KES curve and any data
structure that implements the defined traits.

The channel_opening_protocol test is updated to make use of the
protocols, which now make use of serialising and deserializing the data
structures to simulate communication over the wire.

* Review fixes

* Remove superfluous function