feat: add example using zkuh-rs
docs: example using zkuh-rs

This PR demonstrates how to use the API-only version of nargo and
barretenberg (zkuh-rs).

It replaicates the external NArgo project in `circuits/init` using only
programmatic inputs. This will be NB for security since we don't want to
be storing witnesses at rest where possible.

Performance is also pretty good.

380ms for proving
6ms for verifying

This on an AMD Ryzen 9 7900X.