feat: zk-runner
This PR adds a set of Wrapper objects that facilitate the execution and
verification of Noir circuits in Grease:
- check that the bytecode matches a given checksum, for preventing
  malicious swapping out of circuits.
- a convenient proving wrapper (ProofRunner)
- a convenient verifying wrapper (VerificationRunner) with cached
  verification key to facilitate efficient verifications of the same
types of proof (e.g. after every update). VerificationRunner will also
detect if a proof for a circuit other than the one that is expected is
supplied, since it recalculates the verification key itself.